feat: HighlightEntities now renders entities highlight
