+++ src/components/EntityHighlighter/EntityHighlighter.js
@@ -1,25 +1,10 @@
 import React from 'react';
+import HighlightEntities from './HighlightEntities';
 import EntityActionList from './EntityActionList';
 import NewEntityForm from './NewEntityForm';
-import colors from './colors';
-import { hashString, findEntity, removeEntity } from './helpers';
+import { findEntity, removeEntity } from './helpers';
 
 const styles = {
-  text: {},
-  highlightText: {
-    color: 'transparent',
-    pointerEvents: 'none',
-    padding: '0',
-    whiteSpace: 'pre-wrap',
-    fontFamily: 'source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace',
-    fontSize: 14,
-  },
-  zeroPos: {
-    textAlign: 'left',
-    position: 'absolute',
-    top: 4,
-    left: 2,
-  },
   input: {
     fontFamily: 'source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace',
     fontSize: 14,
@@ -82,21 +67,6 @@ class EntityHighlighter extends React.Component {
     onChange(text, entities);
   }
 
-  renderEntityHighlight = (text, entity, key) => {
-    const start = text.substr(0, entity.start);
-    const end = text.substr(entity.end);
-    const value = text.substr(entity.start, entity.end - entity.start);
-    const color = colors[hashString(entity.label) % colors.length].bg;
-
-    return (
-      <div key={key} style={{ ...styles.zeroPos, ...styles.highlightText }}>
-        <span>{start}</span>
-        <span style={{ opacity: 0.3, backgroundColor: color }}>{value}</span>
-        <span>{end}</span>
-      </div>
-    );
-  };
-
   addEntity = (entityLabel, oldEntity) => {
     const {onChange, text } = this.props;
     let entities = this.props.entities;
@@ -144,7 +114,7 @@ class EntityHighlighter extends React.Component {
             rows={10}
           />
 
-          {entities.map((entity, index) => this.renderEntityHighlight(text, entity, index))}
+          <HighlightEntities text={text} entities={entities} />
         </div>
         <br />
           <NewEntityForm
