+++ src/components/EntityHighlighter/HighlightEntities.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import HighlightEntity from './HighlightEntity';
+
+const EntityActionList = ({
+    text = '',
+    entities = [],
+}) => {
+    return entities.map((entity, index) => <HighlightEntity key={index} text={text} entity={entity} />);
+};
+
+export default EntityActionList;
